dj07/requirements.txt
Django==5.1.2
djangorestframework==3.15.1
telebot==0.0.5
requests==2.32.3
django-cors-headers==4.4.0